Make diffraction_pattern a session wide fixture
